:recycle: [refactor] refactor 뉴앨범, 노티 뷰 #70
