+++ Yeowoo/Sources/CaesarComponents/Component/NotiCardContentsView.swift
@@ -20,7 +20,7 @@ struct NotiCardContentsView: View {
 	var body: some View {
 		if !notis.isParticipateChk {
 			NavigationLink {
-                InvitationView(mainViewModel: mainViewModel, noti: notis, users: users)
+                InvitationView(mainViewModel: mainViewModel, notification: notis, users: users)
 					.navigationBarBackButtonHidden()
 			} label: {
