--- Yeowoo/Sources/CaesarComponents/View/Notification/NotiChildView/NotiView.swift
@@ -1,28 +1,0 @@
-//
-//  NotiView.swift
-//  YeoWooSimpleView
-//
-//  Created by 정회승 on 2023/07/19.
-//
-
-//import SwiftUI
-//
-//struct NotiView: View {
-//
-//    //노티 받기
-//    @State var notication: [SubUser] = [users[1]]
-//
-//    var body: some View {
-//        if notication.isEmpty {
-//            EmptyNotificationView()
-//        } else {
-//            NotificationView()
-//        }
-//    }
-//}
-//
-//struct NotiView_Previews: PreviewProvider {
-//    static var previews: some View {
-//        NotiView()
-//    }
-//}
