+++ Yeowoo/Sources/Views/NewAlbum/FindFriendView.swift
@@ -9,17 +9,18 @@ import SwiftUI
 
 struct FindFriendView: View {
 	
+	@Environment(\.dismiss) var dismiss
+	
 	@EnvironmentObject var appState: AppState
 	
 	@StateObject var viewModel = FindFriendViewModel()
-    @StateObject var mainViewModel: MainViewModel
-	@Environment(\.dismiss) var dismiss
+	@StateObject var mainViewModel: MainViewModel
 	
-	@State var findUser: User = User()
-	@State var friendID: String = ""
-	@State var myFriend: [User] = []
-	@State var selectedFriends: [User] = []
-	@State var friendToggles: [Bool] = []
+	@State private var findUser: User = User()
+	@State private var friendID: String = ""
+	@State private var myFriend: [User] = []
+	@State private var selectedFriends: [User] = []
+	@State private var friendToggles: [Bool] = []
 	@State var newAlbum: Album
 	@State var showAlert = false
 	
@@ -29,7 +30,6 @@ struct FindFriendView: View {
 				.frame(width: UIScreen.width - 50, height: 3)
 				.padding(.top, 15)
 				.foregroundColor(.mainColor)
-			
 			Text("친구 찾기")
 				.font(.system(size: 18, weight: .bold, design: .default))
 				.frame(maxWidth: .infinity, alignment: .leading)
@@ -39,12 +39,10 @@ struct FindFriendView: View {
 			
 			ZStack {
 				GrayTitleMakingView(placeholder: "친구 아이디를 입력해주세요", text: $friendID)
-				
 				HStack {
 					Spacer()
 						.frame(width: UIScreen.width - 110)
 					Button {
-						// 유저 찾기
 						Task {
 							findUser = try await viewModel.searchUser(parmUserId: friendID)
 							if !myFriend.contains(where: { $0.id == findUser.id }) {
@@ -64,7 +62,8 @@ struct FindFriendView: View {
 			
 			ScrollView{
 				ForEach(myFriend.indices, id: \.self) { index in
-					FindFriendContents(user: myFriend[index], friendToggle: $friendToggles[index])
+					FindFriendContents(user: myFriend[index],
+									   friendToggle: $friendToggles[index])
 				}
 				
 				Spacer()
@@ -72,53 +71,35 @@ struct FindFriendView: View {
 			
 			Button{
 				selectedFriends = myFriend.indices.filter { friendToggles[$0] }.map { myFriend[$0] }
-				
-				print("@@@ \(newAlbum.users)")
-				
 				newAlbum.users.append(contentsOf: selectedFriends.map { $0.docId })
-				print("@@@ \(newAlbum.users)")
-				
 				if !newAlbum.users.isEmpty {
 					for _ in 0..<newAlbum.users.count - 1 {
 						newAlbum.role.append("normalFox")
 					}
 				}
-				
-				// 여행 생성
 				Task {
 					try await viewModel.createTravel(newAlbum: newAlbum)
 					showAlert = true
-					
-					// 여기 알림 보내기
-					
-//					let tokens = ["cBQ7l9axn0q7gb3AYab9jf:APA91bGZ3xRKQVgGI5V084qArYZKXON8ypDx_jBqbpNXpQPxgCbJlM-MH3uEZ1eR5LBjyhB063ofEY0QBdpDIgm1k2NY8AcSCx1ZBnc24-xRKDxt0Qz_9GBaWD5H4dftIiBW5bCKU-Zw"]
-					
-//					sendPushNotification(to: tokens, title: "From \(UserDefaultsSetting.nickname)", body: "\(newAlbum.albumTitle)에 초대해요!")
-					
 					sendPushNotification(to: selectedFriends.map { $0.fcmToken }, title: "From \(UserDefaultsSetting.nickname)", body: "\(newAlbum.albumTitle)에 초대해요!")
-					
-//					sendPushNotification(to: [testToken], title: "From \(UserDefaultsSetting.nickname)", body: "\(newAlbum.albumTitle)에 초대해요!")
-					
-				
 				}
 			} label: {
 				Rectangle()
 					.frame(width: UIScreen.main.bounds.width - 30, height: 54)
 					.foregroundColor(Color.mainColor)
 					.cornerRadius(10)
-					.overlay(Text("여행 시작하기").font(.system(size: 18, weight: .bold, design: .default)).foregroundColor(Color.white))
+					.overlay(Text("여행 시작하기").font(.system(size: 18,
+														  weight: .bold,
+														  design: .default)).foregroundColor(Color.white))
 					.padding(.bottom, 20)
 			}
 			.alert(isPresented: $showAlert) {
 				Alert(title: Text("여행 생성 성공"), message: Text("이제 여행을 떠나볼까요?"),
 					  dismissButton: .default(Text("확인"), action: {
-                    Task {
-                        mainViewModel.finishedFetch = false
-                        mainViewModel.openMake.toggle()
-                        try await mainViewModel.loadAlbum()
-                    }
-					// 루트뷰로 이동
-//					self.appState.moveToRootView = true
+					Task {
+						mainViewModel.finishedFetch = false
+						mainViewModel.openMake.toggle()
+						try await mainViewModel.loadAlbum()
+					}
 				}))
 			}
 		}
