+++ Yeowoo/Sources/Views/Role/InvitationChildView/MultiInvitationView.swift
@@ -149,7 +149,7 @@ struct MultiInvitationView: View {
 				}
 				NavigationLink{
 					//참여 로직
-                    RoleSelectView(noti: self.noti, mainViewModel: mainViewModel)
+                    RoleSelectView(notification: self.noti, mainViewModel: mainViewModel)
 						.navigationBarBackButtonHidden()
 					
 				} label: {
